refactor: upload more pre-keys + use pure function
