+++ src/Defaults/index.ts
@@ -62,3 +62,5 @@ export const MEDIA_KEYS = Object.keys(MEDIA_PATH_MAP) as MediaType[]
 export const KEY_BUNDLE_TYPE = Buffer.from([5])
 
 export const MIN_PREKEY_COUNT = 5
+
+export const INITIAL_PREKEY_COUNT = 50
